Load all packages in the root project before resolving their dependencies
Previously, we would load each package and resolve its dependencies
before moving on to the next.  But if one package in the root package
depended on another, we would process that second package as if it was a
vendored package.  This meant that we wouldn't take its test imports
into account.